Fix some cppcheck warnings (format string vs argument mismatches)

+++ vvp/compile.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2001-2013 Stephen Williams (steve@icarus.com)
+ * Copyright (c) 2001-2014 Stephen Williams (steve@icarus.com)
  *
  *    This source code is free software; you can redistribute it
  *    and/or modify it in source code form under the terms of the GNU
@@ -515,16 +515,16 @@ bool vpi_handle_resolv_list_s::resolve(bool mes)
       if (!val.ptr) {
 	    // check for thread vector  T<base,wid>
 	    unsigned base, wid;
-	    unsigned n = 0;
+	    int n = 0;
 	    char ss[32];
 	    if (2 <= sscanf(label(), "T<%u,%u>%n", &base, &wid, &n)
-		&& n == strlen(label())) {
+		&& n == (int)strlen(label())) {
 		  val.ptr = vpip_make_vthr_vector(base, wid, false);
 		  sym_set_value(sym_vpi, label(), val);
 
 	    } else if (3 <= sscanf(label(), "T<%u,%u,%[su]>%n", &base,
 				   &wid, ss, &n)
-		       && n == strlen(label())) {
+		       && n == (int)strlen(label())) {
 
 		  bool signed_flag = false;
 		  for (char*fp = ss ;  *fp ;  fp += 1) switch (*fp) {
@@ -542,13 +542,13 @@ bool vpi_handle_resolv_list_s::resolve(bool mes)
 		  sym_set_value(sym_vpi, label(), val);
 
 	    } else if (2 == sscanf(label(), "W<%u,%[r]>%n", &base, ss, &n)
-		       && n == strlen(label())) {
+		       && n == (int)strlen(label())) {
 
 		  val.ptr = vpip_make_vthr_word(base, ss);
 		  sym_set_value(sym_vpi, label(), val);
 
 	    } else if (1 == sscanf(label(), "S<%u,str>%n", &base, &n)
-		       && n == strlen(label())) {
+		       && n == (int)strlen(label())) {
 
 		  val.ptr = vpip_make_vthr_str_stack(base);
 		  sym_set_value(sym_vpi, label(), val);
